Styling and Company Page Fix
