+++ src/components/Banner/banner.scss
@@ -22,69 +22,6 @@ $green: #90cea1;
             font-size: 2rem;
         }
     }
-    .search{
-        margin-top: 1rem;
-        margin-left: 4rem;
-        width: 70%;
-        display: flex;
-        flex-direction: row;
-        .search-type-wrapper{
-            width: 15%;
-            .search-type{
-                display: flex;
-                flex-direction: row;
-                align-items: center;
-                justify-content: space-evenly;
-                background: $green;
-                padding: 0.5rem;
-                font-size: 1rem;
-                border-radius: 1rem 0 0 1rem;
-                cursor: pointer;
-            }
-            .search-types{
-                z-index: 100;
-                position: absolute;
-                background: white;
-                border-radius: 0.25rem 0.25rem .5rem .5rem;
-                flex-direction: column;
-                display: none;
-                width: 10%;
-                p{
-                    width: 100%;
-                    font-size: 1rem;
-                    text-align: left;
-                    color: #2f2f2f;
-                    padding: .2rem .4rem;
-                    cursor: pointer;
-                    &:hover{
-                        color: $blue;
-                    }
-                }
-            }
-        }
-        .search-input{
-            border: none;
-            width: 70%;
-            background: white;
-            color: #2f2f2f;
-            font-size: 1rem;
-            padding: 0.5rem;
-            outline: none;
-        }
-        .search-btn{
-            width: 10%;
-            border-radius: 0 1rem 1rem 0;
-            background-image: linear-gradient(to right, #90cea1, #01b4e4);
-            color: white;
-            cursor: pointer;
-            padding: 0.5rem;
-            font-size: 1rem;
-            border: none;
-            &:hover{
-               color: #2f2f2f;
-            }
-        }
-    }
 }
 
 .hide{
