+++ src/components/Navbar/Navbar.jsx
@@ -1,24 +1,14 @@
 import { Link } from "react-router-dom";
-import { useNavigate } from "react-router-dom";
-import React, { useState } from "react";
 import Logo from "../../images/tmdb_short.svg";
-import "./navbar.scss";
 import { motion } from "framer-motion";
 import { FiChevronDown } from "react-icons/fi";
-// import { AiOutlineSearch } from "react-icons/ai";
+import { AiOutlineSearch } from "react-icons/ai";
 import { GrLanguage } from "react-icons/gr";
-import { checkType, handleMouseEnter, handleMouseLeave } from "../../shared";
+import { handleMouseEnter, handleMouseLeave } from "../../shared";
+import "./navbar.scss";
+import Search from "../Search/Search";
 
 const Navbar = ({ setLanguage, language }) => {
-  const [input, setInput] = useState("");
-  const [type, setType] = useState("movie");
-  const navigate = useNavigate();
-
-  const submitHandler = (e) => {
-    e.preventDefault();
-    navigate(`/search/${type}/${input}`);
-    setInput("");
-  };
   return (
     <div className="navbar">
       <div className="top-nav">
@@ -119,94 +109,48 @@ const Navbar = ({ setLanguage, language }) => {
         </div>
 
         <div className="top-right-nav">
-          <div className="languages">
-            <GrLanguage size="2em" />
-            {language}
-            <button onClick={() => setLanguage("ar")}>Arabic</button>
-            <button onClick={() => setLanguage("bg")}>Bulgarian</button>
-            <button onClick={() => setLanguage("zh")}>Chinese</button>
-            <button onClick={() => setLanguage("cs")}>Czech</button>
-            <button onClick={() => setLanguage("da")}>Danish</button>
-            <button onClick={() => setLanguage("nl")}>Dutch</button>
-            <button onClick={() => setLanguage("en")}>English</button>
-            <button onClick={() => setLanguage("fr")}>French</button>
-            <button onClick={() => setLanguage("de")}>German</button>
-            <button onClick={() => setLanguage("el")}>Greek</button>
-            <button onClick={() => setLanguage("he")}>Hebrew</button>
-            <button onClick={() => setLanguage("hu")}>Hungarian</button>
-            <button onClick={() => setLanguage("id")}>Indonesian</button>
-            <button onClick={() => setLanguage("it")}>Italian</button>
-            <button onClick={() => setLanguage("ja")}>Japanese</button>
-            <button onClick={() => setLanguage("ko")}>Korean</button>
-            <button onClick={() => setLanguage("pl")}>Polish</button>
-            <button onClick={() => setLanguage("pt")}>Portuguese</button>
-            <button onClick={() => setLanguage("ro")}>Romanian</button>
-            <button onClick={() => setLanguage("ru")}>Russian</button>
-            <button onClick={() => setLanguage("sr")}>Serbian</button>
-            <button onClick={() => setLanguage("es")}>Spanish</button>
-            <button onClick={() => setLanguage("sv")}>Swedish</button>
-            <button onClick={() => setLanguage("tr")}>Turkish</button>
-            <button onClick={() => setLanguage("uk")}>Ukranian</button>
-            <button onClick={() => setLanguage("vi")}>Vietnamese</button>
+          <div className="languages-wrapper">
+            <div className="language-btn">
+              <GrLanguage size="2em" />
+              <p>{language}</p>
+            </div>
+            <div className="languages">
+              <h2>Language Preferences</h2>
+              <button onClick={() => setLanguage("ar")}>Arabic</button>
+              <button onClick={() => setLanguage("bg")}>Bulgarian</button>
+              <button onClick={() => setLanguage("zh")}>Chinese</button>
+              <button onClick={() => setLanguage("cs")}>Czech</button>
+              <button onClick={() => setLanguage("da")}>Danish</button>
+              <button onClick={() => setLanguage("nl")}>Dutch</button>
+              <button onClick={() => setLanguage("en")}>English</button>
+              <button onClick={() => setLanguage("fr")}>French</button>
+              <button onClick={() => setLanguage("de")}>German</button>
+              <button onClick={() => setLanguage("el")}>Greek</button>
+              <button onClick={() => setLanguage("he")}>Hebrew</button>
+              <button onClick={() => setLanguage("hu")}>Hungarian</button>
+              <button onClick={() => setLanguage("id")}>Indonesian</button>
+              <button onClick={() => setLanguage("it")}>Italian</button>
+              <button onClick={() => setLanguage("ja")}>Japanese</button>
+              <button onClick={() => setLanguage("ko")}>Korean</button>
+              <button onClick={() => setLanguage("pl")}>Polish</button>
+              <button onClick={() => setLanguage("pt")}>Portuguese</button>
+              <button onClick={() => setLanguage("ro")}>Romanian</button>
+              <button onClick={() => setLanguage("ru")}>Russian</button>
+              <button onClick={() => setLanguage("sr")}>Serbian</button>
+              <button onClick={() => setLanguage("es")}>Spanish</button>
+              <button onClick={() => setLanguage("sv")}>Swedish</button>
+              <button onClick={() => setLanguage("tr")}>Turkish</button>
+              <button onClick={() => setLanguage("uk")}>Ukranian</button>
+              <button onClick={() => setLanguage("vi")}>Vietnamese</button>
+            </div>
+          </div>
+          <div className="search-btn-wrapper">
+            <AiOutlineSearch size="2em" className="search-btn" />
           </div>
         </div>
       </div>
-
       <div className="bottom-nav">
-        <form onSubmit={submitHandler} className="search">
-          <div
-            className="search-type-wrapper"
-            onMouseEnter={() =>
-              handleMouseEnter("search-options", "search-type-arrow")
-            }
-            onMouseLeave={() =>
-              handleMouseLeave("search-options", "search-type-arrow")
-            }
-          >
-            <div className="search-type">
-              <p>{checkType(type)}</p>
-              <FiChevronDown
-                size="1.2em"
-                id="search-type-arrow"
-                className="arrow"
-              />
-            </div>
-            <div className="search-types" id="search-options">
-              <p
-                onClick={() => setType("movie")}
-                className={type === "movie" ? "hide" : ""}
-              >
-                Movies
-              </p>
-              <p
-                onClick={() => setType("tv")}
-                className={type === "tv" ? "hide" : ""}
-              >
-                Shows
-              </p>
-              <p
-                onClick={() => setType("company")}
-                className={type === "company" ? "hide" : ""}
-              >
-                Company
-              </p>
-              <p
-                onClick={() => setType("person")}
-                className={type === "person" ? "hide" : ""}
-              >
-                Person
-              </p>
-            </div>
-          </div>
-          <input
-            type="text"
-            required
-            onChange={(e) => setInput(e.target.value)}
-            value={input}
-            placeholder={`Search for ${checkType(type)}`}
-          />
-          <button>Search</button>
-        </form>
+        <Search location={"nav"} />
       </div>
     </div>
   );
