+++ src/components/Search/search.scss
@@ -1,0 +1,128 @@
+// Global Variables
+$blue: #01b4e4;
+$green: #90cea1;
+
+.nav-search{
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    background-color: white;
+    border: 1px solid grey;
+
+    .search-type-wrapper{
+        width: 10rem;
+        background-color: $blue;
+        .search-type-icon{
+            height: 100%;
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+            justify-content: space-evenly;
+            p{
+                font-size: 1.2rem;
+            }
+        }
+        .search-types{
+            position: absolute;
+            display: flex;
+            flex-direction: column;
+            text-align: center;
+            background-color: white;
+            color: #2f2f2f;
+            width: 10rem;
+            p{
+                font-size: 1rem;
+                padding: .25rem 0;
+                cursor: pointer;
+                &:hover{
+                    color: $blue;
+                }
+            }
+        }
+    }
+    .nav-input{
+        width: 75%;
+        border: none;
+        color: #2f2f2f;
+        outline: none;
+        padding: .5rem .4rem;
+        font-size: 1rem;
+    }
+
+    .nav-search-btn{
+        width: 10%;
+        background-image: linear-gradient(to right, #90cea1, #01b4e4);
+        color: white;
+        cursor: pointer;
+        padding: 0.5rem;
+        font-size: 1rem;
+        border: none;
+        &:hover{
+            color: #2f2f2f;
+        }
+    }
+}
+
+.search{
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    .search-type-wrapper{
+        width: 15%;
+        .search-type{
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+            justify-content: space-evenly;
+            background: $green;
+            padding: 0.5rem;
+            font-size: 1rem;
+            border-radius: 1rem 0 0 1rem;
+            cursor: pointer;
+        }
+        .search-types{
+            z-index: 100;
+            position: absolute;
+            background: white;
+            border-radius: 0;
+            flex-direction: column;
+            display: none;
+            width: 12%;
+            p{
+                width: 100%;
+                font-size: 1rem;
+                text-align: left;
+                color: #2f2f2f;
+                padding: .2rem .4rem;
+                cursor: pointer;
+                &:hover{
+                    color: $blue;
+                }
+            }
+        }
+    }
+    .search-input{
+        width: 70%;
+        border: none;
+        width: 7;
+        background: white;
+        color: #2f2f2f;
+        font-size: 1rem;
+        padding: 0.5rem;
+        outline: none;
+    }
+    .search-btn{
+        width: 15%;
+        border-radius: 0 1rem 1rem 0;
+        background-image: linear-gradient(to right, #90cea1, #01b4e4);
+        color: white;
+        cursor: pointer;
+        padding: 0.5rem;
+        font-size: 1rem;
+        border: none;
+        &:hover{
+           color: #2f2f2f;
+        }
+    }
+}
